+++ src/modules/plugins/plugins.service.ts
@@ -844,17 +844,6 @@ export class PluginsService {
     // Modify this plugins schema to set the default port number
     if (pluginName === this.configService.name) {
       configSchema.schema.properties.port.default = this.configService.ui.port
-
-      // Filter some options from the UI config when using service mode
-      if (this.configService.serviceMode) {
-        configSchema.layout = configSchema.layout.filter((x: any) => {
-          return x.ref !== 'log'
-        })
-
-        configSchema.layout = configSchema.layout.filter((x: any) => {
-          return !(x === 'sudo' || x.key === 'restart')
-        })
-      }
     }
 
     // Modify homebridge-alexa to set the default pin