+++ ui/src/app/modules/status/widgets/system-info-widget/system-info-widget.component.ts
@@ -62,16 +62,4 @@ export class SystemInfoWidgetComponent implements OnInit {
     ref.componentInstance.faIconClass = 'fab fa-fw fa-node-js primary-text'
     ref.componentInstance.ctaButtonLink = 'https://github.com/homebridge/homebridge-config-ui-x/wiki/Troubleshooting/#error---update-node---your-version-of-linux-does-not-meet-the-glibc-version'
   }
-
-  serviceModeModal() {
-    const ref = this.$modal.open(InformationComponent, {
-      size: 'lg',
-      backdrop: 'static',
-    })
-    ref.componentInstance.title = this.$translate.instant('status.widget.info.servicemode_title')
-    ref.componentInstance.message = this.$translate.instant('status.widget.info.servicemode_message')
-    ref.componentInstance.ctaButtonLabel = this.$translate.instant('form.button_more_info')
-    ref.componentInstance.faIconClass = 'fas fa-fw fa-circle-exclamation primary-text'
-    ref.componentInstance.ctaButtonLink = 'https://github.com/homebridge/homebridge-config-ui-x/wiki/How-To-Swap-From-Standalone-Mode-to-Service-Mode'
-  }
 }